refactor: extract home screen logic into hooks
